# openapi.yaml
openapi: 3.0.3
info:
  title: Environmental Monitoring API
  version: 1.0.0
  description: |
    Production-grade REST API for environmental monitoring, mission management,
    visitor tracking, and geospatial analytics.

    Scenario-scoped routes are prefixed with:
    /api/v1/{scenario_name}/...
servers:
  - url: http://localhost:8000
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/scenarios:
    get:
      tags: [Scenarios]
      summary: List scenarios
      responses:
        '200':
          description: Scenarios list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScenarioResponse'
    post:
      tags: [Scenarios]
      summary: Create a scenario (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioCreate'
      responses:
        '201':
          description: Scenario created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/scenarios/{scenario_name}:
    get:
      tags: [Scenarios]
      summary: Get scenario by name
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      responses:
        '200':
          description: Scenario details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScenarioResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/{scenario_name}/visitors:
    get:
      tags: [Visitors]
      summary: List visitors
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Paginated visitors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVisitors'
    post:
      tags: [Visitors]
      summary: Create visitor
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitorCreate'
      responses:
        '201':
          description: Visitor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorResponse'
  /api/v1/{scenario_name}/visitors/{visitor_id}:
    get:
      tags: [Visitors]
      summary: Get visitor by id
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/VisitorId'
      responses:
        '200':
          description: Visitor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/{scenario_name}/visitors/{visitor_id}/body-measurements:
    get:
      tags: [Visitors]
      summary: List body measurements for visitor
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/VisitorId'
      responses:
        '200':
          description: Body measurements list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BodyMeasurementResponse'
    post:
      tags: [Visitors]
      summary: Create body measurement for visitor
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/VisitorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyMeasurementCreate'
      responses:
        '201':
          description: Body measurement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BodyMeasurementResponse'
  /api/v1/{scenario_name}/visitors/{visitor_id}/tracks:
    get:
      tags: [Visitor Tracks]
      summary: List tracks for visitor
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/VisitorId'
      responses:
        '200':
          description: Visitor tracks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisitorTrackResponse'
    post:
      tags: [Visitor Tracks]
      summary: Create track for visitor
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/VisitorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitorTrackCreate'
      responses:
        '201':
          description: Track created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitorTrackResponse'
  /api/v1/{scenario_name}/visitor-tracks:
    get:
      tags: [Visitor Tracks]
      summary: List all visitor tracks in scenario
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/Page'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Paginated visitor tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedVisitorTracks'
  /api/v1/{scenario_name}/missions:
    get:
      tags: [Missions]
      summary: List missions
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [planned, active, completed, cancelled]
      responses:
        '200':
          description: Paginated missions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedMissions'
    post:
      tags: [Missions]
      summary: Create mission
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionCreate'
      responses:
        '201':
          description: Mission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionResponse'
  /api/v1/{scenario_name}/missions/{mission_id}:
    get:
      tags: [Missions]
      summary: Get mission by id
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MissionId'
      responses:
        '200':
          description: Mission details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Missions]
      summary: Update mission
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MissionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionUpdate'
      responses:
        '200':
          description: Mission updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionResponse'
    delete:
      tags: [Missions]
      summary: Delete mission
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MissionId'
      responses:
        '204':
          description: Mission deleted
  /api/v1/{scenario_name}/missions/{mission_id}/assignments:
    post:
      tags: [Missions]
      summary: Assign instrument to mission
      parameters:
        - $ref: '#/components/parameters/ScenarioName'
        - $ref: '#/components/parameters/MissionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionAssignmentCreate'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionAssignmentResponse'
...
components:
  parameters:
    ScenarioName:
      name: scenario_name
      in: path
      required: true
      schema:
        type: string
      description: Scenario identifier/name, e.g. drill-2026
...
    VisitorId:
      name: visitor_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MissionId:
      name: mission_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict / duplicate resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
...
    ScenarioCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        type:
          type: string
          enum: [drill, real_event, training, other]
          default: drill
        description:
          type: string
          nullable: true
          maxLength: 500
    ScenarioResponse:
      type: object
      required: [id, name, type, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [drill, real_event, training, other]
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    VisitorCreate:
      type: object
      properties:
        demographics:
          type: object
          additionalProperties: true
        tags:
          type: array
          nullable: true
          items: { type: string }
    VisitorResponse:
      type: object
      required: [id, scenario_id, created_at, demographics]
      properties:
        id: { type: string, format: uuid }
        scenario_id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        demographics:
          type: object
          additionalProperties: true
        tags:
          type: array
          nullable: true
          items: { type: string }
    BodyMeasurementCreate:
      type: object
      required: [timestamp, type, value, unit]
      properties:
        timestamp: { type: string, format: date-time }
        type: { type: string, maxLength: 100 }
        value: { type: number, format: float }
        unit: { type: string, maxLength: 50 }
    BodyMeasurementResponse:
      type: object
      required: [id, visitor_id, timestamp, type, value, unit]
      properties:
        id: { type: string, format: uuid }
        visitor_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        type: { type: string }
        value: { type: number, format: float }
        unit: { type: string }
    VisitorTrackCreate:
      type: object
      required: [geom]
      properties:
        geom:
          $ref: '#/components/schemas/GeoJSONLineString'
    VisitorTrackResponse:
      type: object
      required: [id, visitor_id, recorded_at]
      properties:
        id: { type: string, format: uuid }
        visitor_id: { type: string, format: uuid }
        geom:
          $ref: '#/components/schemas/GeoJSONLineString'
        recorded_at: { type: string, format: date-time }
    MissionCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 200 }
        description:
          type: string
          nullable: true
          maxLength: 1000
        status:
          type: string
          enum: [planned, active, completed, cancelled]
          default: planned
        target_area:
          $ref: '#/components/schemas/GeoJSONPolygon'
    MissionUpdate:
      type: object
      properties:
        name: { type: string, maxLength: 200 }
        description:
          type: string
          nullable: true
          maxLength: 1000
        status:
          type: string
          enum: [planned, active, completed, cancelled]
        target_area:
          $ref: '#/components/schemas/GeoJSONPolygon'
    MissionResponse:
      type: object
      required: [id, scenario_id, name, status, created_at]
      properties:
        id: { type: string, format: uuid }
        scenario_id: { type: string, format: uuid }
        name: { type: string }
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [planned, active, completed, cancelled]
        target_area:
          $ref: '#/components/schemas/GeoJSONPolygon'
        created_at: { type: string, format: date-time }
    MissionAssignmentCreate:
      type: object
      required: [instrument_id]
      properties:
        instrument_id: { type: string, format: uuid }
    MissionAssignmentResponse:
      type: object
      required: [id, mission_id, instrument_id, assigned_at]
      properties:
        id: { type: string, format: uuid }
        mission_id: { type: string, format: uuid }
        instrument_id: { type: string, format: uuid }
        assigned_at: { type: string, format: date-time }
...
    GeoJSONLineString:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [LineString]
        coordinates:
          type: array
          minItems: 2
          items:
            type: array
            minItems: 2
            items:
              type: number
              format: double
    GeoJSONPolygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              minItems: 2
              items:
                type: number
                format: double
...
    PaginatedVisitors:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VisitorResponse'
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }
    PaginatedMissions:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MissionResponse'
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }
    PaginatedVisitorTracks:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/VisitorTrackResponse'
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        pages: { type: integer }
...